Rename ExecutionTrace filename
