+++ libethploit/CallTrace.cpp
@@ -1,4 +1,4 @@
-#include "ExecutionTrace.h"
+#include "CallTrace.h"
 
 namespace dev 
 {
